requirements.txt
PyMuPDF>=1.23.8
anthropic>=0.8.1
Pillow>=10.2.0
PyMuPDF>=1.23.8
httpx==0.27.2